+++ ui/src/shared/components/time_series/TimeSeries.tsx
@@ -128,6 +128,20 @@ class TimeSeries extends PureComponent<Props, State> {
     const currQueries = _.map(this.props.queries, q => q.text)
     const queriesDifferent = !_.isEqual(prevQueries, currQueries)
 
+    let manualSubmit = false
+    if (!queriesDifferent) {
+      for (let i = 0; i < this.props.queries.length; i++) {
+        const query = this.props.queries[i]
+        const prevQuery = prevProps.queries[i]
+        if (
+          query.queryConfig?.status?.isManuallySubmitted &&
+          !prevQuery.queryConfig?.status?.isManuallySubmitted
+        ) {
+          manualSubmit = true
+          break
+        }
+      }
+    }
     const prevTemplates = _.get(prevProps, 'templates')
     const newTemplates = _.get(this.props, 'templates')
     // templates includes dashTime and upperDashTime which capture zoomedTimeRange
@@ -140,6 +154,7 @@ class TimeSeries extends PureComponent<Props, State> {
     const timeRangeChanged = oldLower !== newLower || oldUpper !== newUpper
 
     const shouldExecuteQueries =
+      manualSubmit ||
       queriesDifferent ||
       timeRangeChanged ||
       templatesDifferent ||
@@ -308,7 +323,13 @@ class TimeSeries extends PureComponent<Props, State> {
     const {source, templates, editQueryStatus, queries} = this.props
 
     for (const query of queries) {
-      editQueryStatus(query.id, {loading: true})
+      const prevStatus = query.queryConfig.status
+      editQueryStatus(query.id, {
+        loading: true,
+        isManuallySubmitted: prevStatus?.isManuallySubmitted,
+        submittedStatus: prevStatus?.submittedStatus,
+        submittedQuery: prevStatus?.submittedQuery,
+      })
     }
 
     const results = await this.executeInfluxQLQueries(
@@ -335,8 +356,18 @@ class TimeSeries extends PureComponent<Props, State> {
           queryStatus = {success: 'Success!'}
         }
       }
-
-      editQueryStatus(query.id, queryStatus)
+      const shouldPreserve =
+        query.queryConfig.isExcluded &&
+        !query.queryConfig.status?.isManuallySubmitted
+      editQueryStatus(query.id, {
+        ...queryStatus,
+        submittedStatus: shouldPreserve
+          ? query.queryConfig.status.submittedStatus
+          : queryStatus,
+        submittedQuery: shouldPreserve
+          ? query.queryConfig.status.submittedQuery
+          : query.text,
+      })
     }
 
     const validQueryResults = results