+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [unreleased]
 
+### Bug Fixes
+
+1. [#6103](https://github.com/influxdata/chronograf/pull/6103): Set active database for InfluxQL meta queries.
+
 ### Other
 
 1. [#6102](https://github.com/influxdata/chronograf/pull/6102): Upgrade golang to 1.21.12.